backend/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
psycopg2   # psycopg3 (Python 3.13 호환)
pandas
numpy
pyarrow  # Parquet 파일 지원
redis
